+++ Source/SFConditional/Private/SFConditionalAnswer.cpp
@@ -54,6 +54,10 @@ FSFConditionalAnswer SF::Conditional::Answer::FromErrorMsg(const FName& InErrorM
 	static FSFConditionalAnswer Instance{ FName(ErrorMsg) }; \
 return Instance;
 
+#define SF_CONDITIONAL_IMPL_ERROR_OBJ(ErrorMsg, Object) \
+	static FSFConditionalAnswer Instance{ FName(FString::Printf(TEXT(ErrorMsg), IsValid(Object) ? *Object->GetName() : *FString("nullptr"))) }; \
+return Instance;
+
 const FSFConditionalAnswer& SF::Conditional::Answer::Error::InvalidTestObject()
 {
 	SF_CONDITIONAL_IMPL_ERROR("Invalid test object");
@@ -74,64 +78,64 @@ const FSFConditionalAnswer& SF::Conditional::Answer::Error::HasChildWithRuntimeE
 	SF_CONDITIONAL_IMPL_ERROR("Has child with runtime error");
 }
 
-const FSFConditionalAnswer& SF::Conditional::Answer::Error::NoPlayerController()
+const FSFConditionalAnswer& SF::Conditional::Answer::Error::NoPlayerController(const UWorld* World)
 {
-	SF_CONDITIONAL_IMPL_ERROR("Could not find valid player controller");
+	SF_CONDITIONAL_IMPL_ERROR_OBJ("Could not find valid player controller in %s", World);
 }
 
-const FSFConditionalAnswer& SF::Conditional::Answer::Error::NoViewport()
+const FSFConditionalAnswer& SF::Conditional::Answer::Error::NoViewport(const APlayerController* Pc)
 {
-	SF_CONDITIONAL_IMPL_ERROR("Could not find valid viewport");
+	SF_CONDITIONAL_IMPL_ERROR_OBJ("Could not find valid viewport for %s", Pc);
 }
 
-const FSFConditionalAnswer& SF::Conditional::Answer::Error::Instigator::NoActor()
+const FSFConditionalAnswer& SF::Conditional::Answer::Error::Instigator::NoActor(const UObject* TestObject)
 {
-	SF_CONDITIONAL_IMPL_ERROR("Could not resolve instigator to supported class AActor");
+	SF_CONDITIONAL_IMPL_ERROR_OBJ("Could not resolve instigator to supported class AActor: %s", TestObject);
 }
 
-const FSFConditionalAnswer& SF::Conditional::Answer::Error::Instigator::NoActorComponent()
+const FSFConditionalAnswer& SF::Conditional::Answer::Error::Instigator::NoActorComponent(const UObject* TestObject)
 {
-	SF_CONDITIONAL_IMPL_ERROR("Could not resolve instigator to supported class UActorComponent");
+	SF_CONDITIONAL_IMPL_ERROR_OBJ("Could not resolve instigator to supported class UActorComponent: %s", TestObject);
 }
 
-const FSFConditionalAnswer& SF::Conditional::Answer::Error::Instigator::NoSceneComponent()
+const FSFConditionalAnswer& SF::Conditional::Answer::Error::Instigator::NoSceneComponent(const UObject* TestObject)
 {
-	SF_CONDITIONAL_IMPL_ERROR("Could not resolve instigator to supported class USceneComponent");
+	SF_CONDITIONAL_IMPL_ERROR_OBJ("Could not resolve instigator to supported class USceneComponent: %s", TestObject);
 }
 
-const FSFConditionalAnswer& SF::Conditional::Answer::Error::Instigator::NoActorProvider()
+const FSFConditionalAnswer& SF::Conditional::Answer::Error::Instigator::NoActorProvider(const UObject* TestObject)
 {
-	SF_CONDITIONAL_IMPL_ERROR("Could not resolve instigator to supported class UActorComponent");
+	SF_CONDITIONAL_IMPL_ERROR_OBJ("Could not resolve instigator to supported class UActorComponent: %s", TestObject);
 }
 
-const FSFConditionalAnswer& SF::Conditional::Answer::Error::Instigator::NoTransformProvider()
+const FSFConditionalAnswer& SF::Conditional::Answer::Error::Instigator::NoTransformProvider(const UObject* TestObject)
 {
-	SF_CONDITIONAL_IMPL_ERROR("Could not resolve instigator to a FTransform");
+	SF_CONDITIONAL_IMPL_ERROR_OBJ("Could not resolve instigator to a FTransform: %s", TestObject);
 }
 
-const FSFConditionalAnswer& SF::Conditional::Answer::Error::TestObject::NoActor()
+const FSFConditionalAnswer& SF::Conditional::Answer::Error::TestObject::NoActor(const UObject* TestObject)
 {
-	SF_CONDITIONAL_IMPL_ERROR("Could not resolve test object to supported class AActor");
+	SF_CONDITIONAL_IMPL_ERROR_OBJ("Could not resolve test object to supported class AActor: %s", TestObject);
 }
 
-const FSFConditionalAnswer& SF::Conditional::Answer::Error::TestObject::NoActorComponent()
+const FSFConditionalAnswer& SF::Conditional::Answer::Error::TestObject::NoActorComponent(const UObject* TestObject)
 {
-	SF_CONDITIONAL_IMPL_ERROR("Could not resolve test object to supported class UActorComponent");
+	SF_CONDITIONAL_IMPL_ERROR_OBJ("Could not resolve test object to supported class UActorComponent: %s", TestObject);
 }
 
-const FSFConditionalAnswer& SF::Conditional::Answer::Error::TestObject::NoSceneComponent()
+const FSFConditionalAnswer& SF::Conditional::Answer::Error::TestObject::NoSceneComponent(const UObject* TestObject)
 {
-	SF_CONDITIONAL_IMPL_ERROR("Could not resolve test object to supported class USceneComponent");
+	SF_CONDITIONAL_IMPL_ERROR_OBJ("Could not resolve test object to supported class USceneComponent: %s", TestObject);
 }
 
-const FSFConditionalAnswer& SF::Conditional::Answer::Error::TestObject::NoActorProvider()
+const FSFConditionalAnswer& SF::Conditional::Answer::Error::TestObject::NoActorProvider(const UObject* TestObject)
 {
-	SF_CONDITIONAL_IMPL_ERROR("Could not resolve test object to supported class AActor or UActorComponent");
+	SF_CONDITIONAL_IMPL_ERROR_OBJ("Could not resolve test object to supported class AActor or UActorComponent: %s", TestObject);
 }
 
-const FSFConditionalAnswer& SF::Conditional::Answer::Error::TestObject::NoTransformProvider()
+const FSFConditionalAnswer& SF::Conditional::Answer::Error::TestObject::NoTransformProvider(const UObject* TestObject)
 {
-	SF_CONDITIONAL_IMPL_ERROR("Could not resolve test object to a FTransform");
+	SF_CONDITIONAL_IMPL_ERROR_OBJ("Could not resolve test object to a FTransform: %s", TestObject);
 }
 
 #undef SF_CONDITIONAL_IMPL_ERROR