+++ Source/SFConditionalTests/Private/Specs/ConditionalTypes/SFConditionalTypeActorIsTicking.spec.cpp
@@ -30,7 +30,7 @@ void FConditionalTypesActorIsTickingSpec::Define()
 	{
 		It("should yield NoActor error state", [this]
 		{
-			TestEqual("Conditional Answer", Sut->EvaluateObject(Object), Answer::Error::TestObject::NoActor());
+			TestEqual("Conditional Answer", Sut->EvaluateObject(Object), Answer::Error::TestObject::NoActor(Object));
 		});
 	});